+++ lib/list_route.dart
@@ -1,0 +1,69 @@
+import 'dart:convert';
+
+import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
+
+class ListRoute extends StatefulWidget {
+  ListRoute({Key key}) : super(key: key);
+
+  @override
+  _ListRouteState createState() => _ListRouteState();
+}
+
+class _ListRouteState extends State<ListRoute> {
+  List widgets = [];
+
+  @override
+  void initState() {
+    super.initState();
+    loadData();
+  }
+
+  Future<void> loadData() async {
+    var dataURL = Uri.parse('https://jsonplaceholder.typicode.com/posts');
+    http.Response response = await http.get(dataURL);
+    setState(() {
+      widgets = jsonDecode(response.body);
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Planowanie'),
+      ),
+      body: ListView.builder(
+        itemCount: widgets.length,
+        itemBuilder: (BuildContext context, int position) {
+          return getRow(position);
+          // todo how to do getRow(widgets[position]) to extract data from UI? How to know what's the type of widgets[position]?
+        },
+      ),
+    );
+  }
+
+  // todo extract to SingleListWidget widget. How?
+  Widget getRow(int i) {
+    return MaterialButton(
+        onPressed: () {
+          Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => DetailsRoute()),
+          );
+        },
+        child: Padding(
+          padding: EdgeInsets.all(12.0),
+          child: Text("Row ${widgets[i]["title"]}"),
+        ));
+  }
+}
+
+class DetailsRoute extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text("DETAILS PAGE"),
+    );
+  }
+}
